refactor(experiments): extracting experiments and shared metrics common functions into modules.
