+++ frontend/src/scenes/experiments/activity-descriptions/index.ts
@@ -1,0 +1,2 @@
+export * from './experimentChangeDescription'
+export * from './sharedMetricChangeDescription'
